refactor(frontend): decouple productivity screen UI and logic
- Extract state management and data filtering to `useProductivityViewModel`
- Extract inline render item for events to `EventCard` component
- Extract list empty component to `ProductivityEmptyState` component
- Adhere to the Refactoring Engineer constraints: Decompose, Decouple, Clarify, Harden
