+++ frontend/src/components/productivity/EventCard.tsx
@@ -1,0 +1,90 @@
+import React from 'react';
+import { StyleSheet, View } from 'react-native';
+import { Ionicons } from '@expo/vector-icons';
+import { useTranslation } from 'react-i18next';
+import { AppText } from '../AppText';
+import { CalendarEvent } from '@/core/models';
+import { theme } from '@/core/theme';
+import { DESIGN_TOKENS } from '@/core/design/tokens';
+
+const T = {
+  surface: { light: DESIGN_TOKENS.colors.surface },
+  border: { light: DESIGN_TOKENS.colors.border },
+  onSurface: {
+    light: DESIGN_TOKENS.colors.text,
+    mutedLight: DESIGN_TOKENS.colors.muted,
+  },
+  primary: {
+    DEFAULT: DESIGN_TOKENS.colors.primary,
+    surfaceLight: DESIGN_TOKENS.colors.primarySoft,
+  },
+};
+const S = theme.spacing;
+const R = theme.borderRadius;
+
+interface EventCardProps {
+  event: CalendarEvent;
+}
+
+/**
+ * A component to display a calendar event in the productivity lists.
+ */
+export function EventCard({ event }: EventCardProps) {
+  const { i18n } = useTranslation();
+
+  return (
+    <View style={styles.eventCard}>
+      <View style={styles.eventIcon}>
+        <Ionicons name="calendar-outline" size={16} color={T.primary.DEFAULT} />
+      </View>
+      <View style={styles.eventBody}>
+        <AppText style={styles.eventTitle}>{event.title}</AppText>
+        <AppText style={styles.eventMeta}>
+          {new Intl.DateTimeFormat(i18n.language, {
+            weekday: 'short',
+            month: 'short',
+            day: 'numeric',
+            hour: event.is_all_day ? undefined : '2-digit',
+            minute: event.is_all_day ? undefined : '2-digit',
+          }).format(new Date(event.start_time))}
+        </AppText>
+      </View>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  eventCard: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: T.surface.light,
+    borderWidth: 1,
+    borderColor: T.border.light,
+    borderRadius: R.xl,
+    padding: S.lg,
+    marginBottom: S.md,
+    ...theme.elevation.sm,
+  },
+  eventIcon: {
+    width: 32,
+    height: 32,
+    borderRadius: R.lg,
+    backgroundColor: T.primary.surfaceLight,
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginRight: S.md,
+  },
+  eventBody: {
+    flex: 1,
+  },
+  eventTitle: {
+    color: T.onSurface.light,
+    fontWeight: '700',
+    fontSize: 15,
+  },
+  eventMeta: {
+    color: T.onSurface.mutedLight,
+    marginTop: 2,
+    fontSize: 13,
+  },
+});
